sites/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
cffi==1.5.0
characteristic==14.3.0
cryptography==1.2.2
cssselect==0.9.1
decorator==4.0.6
Django==1.9
django-celery==3.1.17
django-dynamic-scraper==0.10.0
django-haystack==2.4.1
enum34==1.1.2
future==0.15.2
idna==2.0
ipaddress==1.0.16
jsonpath-rw==1.4.0
kombu==3.0.33
lxml==3.5.0
Pillow==3.1.0
ply==3.8
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
PyPDF2==1.25.1
pysolr==3.4.0
pytz==2015.7
queuelib==1.4.2
requests==2.9.1
Scrapy==1.0.4
scrapy-djangoitem==1.0.0
scrapyd==1.1.0
scrapyjs==0.1.1
service-identity==14.0.0
six==1.10.0
Twisted==15.5.0
w3lib==1.13.0
wheel==0.26.0
zope.interface==4.1.3
